fix(api): cap email retry backoff exponent and use NUMERIC for volume aggregation
Fixes #1126: mark_failed computed 2_u64.pow(new_attempts) with no cap on the
exponent. If max_attempts is ever misconfigured to a large value, attempts
climbing past ~58 overflow the pow, panicking in debug builds or wrapping to
a bogus retry time in release builds. Extracted the backoff computation into
EmailQueue::compute_backoff_seconds, which clamps the exponent to
MAX_BACKOFF_EXPONENT (20) before the pow, and added tests asserting the
computation never overflows/wraps and stays monotonically increasing up to
the cap.

Fixes #1127: Statistics.total_volume and FeaturedMarket.volume were
aggregated via SUM() on a DOUBLE PRECISION column and exposed as f64,
inconsistent with the exact-string convention used for on-chain volumes
elsewhere in the API. Added migration 021 converting markets.total_volume to
NUMERIC, changed the aggregation queries to cast the result to TEXT, and
changed Statistics.total_volume / FeaturedMarket.volume / FeaturedMarketView.volume
to String. Verified against a live Postgres 16 instance that summing 0.1 and
0.2 now yields exactly "0.3" instead of the float-rounded 0.30000000000000004.
Also fixed the featured_markets_cached ORDER BY, which would otherwise sort
by the new TEXT-aliased column lexicographically instead of numerically.

Updated openapi.yaml and database/schema.sql to match.

+++ services/api/database/schema.sql
@@ -242,13 +242,13 @@ CREATE TABLE IF NOT EXISTS public.audit_log (
     CONSTRAINT audit_log_pkey PRIMARY KEY (id)
 );
 
--- migration 011
+-- migration 011, 021
 CREATE TABLE IF NOT EXISTS public.markets (
     id            bigint           NOT NULL GENERATED ALWAYS AS IDENTITY,
     title         text             NOT NULL,
     status        text             NOT NULL DEFAULT 'active',
     outcome_index integer,
-    total_volume  double precision NOT NULL DEFAULT 0,
+    total_volume  numeric          NOT NULL DEFAULT 0,
     ends_at       timestamptz      NOT NULL,
     created_at    timestamptz      NOT NULL DEFAULT now(),
     resolved_at   timestamptz,
